ch08/django_sse_chat/requirements.txt
openai==2.21.0
python-dotenv==1.2.1
Pillow==11.1.0
django==5.2.11